dependencies: Handle .la lib files returned by pkgconfig in uninstalled setups
When building against software that is being built uninstalled,
pkg-config returns values from the -uninstalled.pc variant which
might contain .la files as --libs.

This patch opens the .la files to figure out where the actual shared
library are.

Some part of this is inspired by what is done in the
gobject-introspection giscanner/utils.py code