Fix cross linking
I could not get linking to work using gcc so this adds code to add
support for a linkerlist to the GnuCCompiler.

Also the I couldn't get the linker script, link.meson.ld, to work
when the executable target directory had an @ in the name, so the
change to build.py changes the @ to '-'.