Change type_suffix from @xxx to -xxx.
Using an '@' in a path in linker scripts is not allowed, and I can
imagine it causing problems with other programs. So I suggest
using dash, '-', rather than '@' in the type_suffix() methods.

As an example in baremetal-hi/link.meson.ld, the current solution that
Jussi provided was to use an asterisk, '*', in the file name see:
winksaville/baremetal-hi@d358e06

But using an asterisk to resolve the problem could cause multiple matches
leading to unforeseen errors.