Replace c++1y with c++14 as a choice for cpp_std
c++1y is deprecated in GCC 5 (see -std option in man page). Not sure about
Clang since the documentation is a bit more sparse. But it is probably a
good idea to keep them in sync and also to prefer c++14 over c++1y since
c++1y was only temporary until the standard was done.