compilers: Pass --no-undefined to GNU ld by default
This is the default for all other linkers, and most projects already set this by
default. The only reason to not have this is obscure platforms like BeOS and
HPPA that use this as a feature. For those, the required symbols should be
explicitly set as undefined with the --undefined linker option.

To disable this, pass c_no_undefined=off and cpp_no_undefined=off as compiler
options or default project options.