validate all options before generating build files
This fails the build when passing unknown options via command line or as
'default_options' in meson.build.

It prints all unknown project options that were used together with a list
of all known options, including those from subprojects.

It also prints a warning for all options that are not in use for the
current build, but might be on another platform (currently only compiler
options that might exist on other platforms like when using 'cpp_eh' with
gcc on Linux).