VSCompiler: Instead of using /FS, set /Fd$pdbfilename
/FS forces .pdb file writing to be serialized through MSPDBSRV.EXE,
which solves most of the concurrency issues (esp. with anti-viruses),
but doesn't solve all of them. I was able to reproduce an intermittent
'cannot open program database vc140.pdb'. vcXX0.pdb is the default
temporary filename for the pdb file if none is specified. In many cases
the pdb file not being found was silently causing debug symbols to not
be generated.

By setting /Fd, we speed up debug builds and avoid concurrency issues.
I haven't tested with an antivirus besides Windows Defender yet, but it
should work because every target has its own pdb file now. Note that the
correct solution is to add your development directory to the list of
excluded paths otherwise the AV will use up a large chunk of your CPU
resources scanning every single generated file.

We just always set it because it's excessively invasive w.r.t. layering
to only set it when we're in a debug mode, and the compiler will ignore
/Fd anyway if /Zi or /ZI aren't specified.