vs backends: Translate unix link and compile flags
We allow build files to contain (at least) -l and -L link flags, and we
filter out MinGW-specific compiler flags that are not valid or needed
with MSVC such as -mms-bitfields. So convert those to msvc-style
arguments in the same way that the Ninja backend does.

Global arguments, build type arguments, etc which are always
compiler-specific don't need to be translated.