vs: Never serialize debug file generation with /FS
/FS forces .pdb file writing to be serialized through MSPDBSRV.EXE,
which is supposed to solve concurrency issues (esp. with anti-viruses),
but it doesn't actually always work. It also isn't needed anymore since
we don't use a shared vcXXX.pdb temporary file for pdb writing and use
/Fd to explicitly set a per-target pdb filename instead.

Not serializing the PDB writing will make a large difference on fast
disks when linking multiple libraries and executables simultaneously.