New API: find_python3 returns full path to Python3
The executable name for Python varies depending on the platform that
you're on. It can be 'python' (Windows, Gentoo, Arch Linux, etc),
'python3' (Debian, Fedora, Ubuntu), 'python3.5' (NetBSD, etc), and so
on. Since Meson is written in Python 3, we always know what the full
path to the Python 3 executable is, so this makes build files much more
portable.

find_python2 can be added if there is demand for it, but ideally scripts
used for building should just use Python 3 to minimize build
dependencies.