Add string '.matches' method to match glob patterns
This allows to write some if statements with lots of options
in a tighter way, or to replace most of the common case/esac
blocks used in configure scripts, e.g.

if host_cpu_family.matches('x86|x86_64|s390?|arm*|crisv32*|etrax*')
 ...
endif

Special characters for shell-style wildcards supported are:

 *      matches everything
 ?      matches a single character
 [seq]  matches any character in seq, e.g. [12345] or [1-5] or [a-z]
 [!seq] matches any character not in seq

Multiple expressions can be checked for in one go with '|' as OR operator.