compilers: Consolidate language suffix mapping
Instead of spreading it across the entire file and re-implementing
can_compile for each compiler, consolidate it in one dictionary which
the base compiler class can refer to based on self.language set by the
child class.

This has fixed a number of hidden and unreported bugs due to bitrotten
can_compiler implementations.