gnome.py: compile gschemas to share/glib-2.0/schemas/
Currently the compile_schemas meson tool only works if the source of
the schemas is in the same directory as from where it is called, and
also doesn't respect the standard install directory for schemas, which
is $prefix/share/glib-2.0/schemas.

Make compile_schemas accept a positional parameter for the source
directory and compile the output to $prefix/share/glib-2.0/schemas.