dependencies: Use a wrapper for calling pkg-config
Reduces duplicated code, and also use universal_newlines=True which
avoids having to decode the bytes output to string.

We mostly need this so we can pass the *current* process environment to
pkg-config which might've changed since the module was imported. Without
this, subprocess.Popen invokes pkg-config with a stale environment (used
in the unittest added later)