gnome: Add check_gir() method
Autotools Archive recently gained macros for checking for
gobject-introspection runtime dependencies[0]. While those
are not required to build a project, a clear error message
at build-time is much easier for users or packagers to figure
out than possibly cryptic runtime failures, so those checks
turn out to be quite useful for non-compiled languages.
To not lose that ability when moving to meson, add a check_gir()
method to the gnome module that implements the functionality of
the AX_CHECK_GIRS_GJS and AC_CHECK_GIR_SYMBOLS_GJS macros.

[0] http://endlessmblog.com/gjs-and-autoconf/