Hotfix for cross-compilation from Windows to Linux
We currently pass cross-compiler options to the native compiler too and
when cross-compiling from Windows to Linux, `options` will contain
Linux-specific options which doesn't include `c_winlibs`.

The proper fix is to allow cross-info files to specify compiler options
and to maintain both cross and native compiler options in coredata, but
that will have to be done after the 0.36.0 release.

Also fixes a typo in MinGW cpp_winlibs option setting.

Closes #1029