qt: Fix qmake detection on Linux and Windows
qmake for both Qt4 and Qt5 detection was assuming that it was only used
on Windows, which is incorrect. It can also be used on Linux for
cross-compilation or in general when pkg-config is not available.

It was also not failing properly for both Qt5 and Qt4 when no libraries
were found, and was assuming that the .dll was always available.

Qt4 detection with qmake was also completely broken.

Also prevents unwanted injection of partially-found qt dependencies in
targets by unsetting self.cargs and self.largs