mesontest: Let user work with gdb when not running in a loop
In the normal case the user probably wants to set break point or
anything when running an app in gdb, we should let him a chance to
do so.

In the case he is running in a loop, it probably means he want to
reproduce a crash or a race inside gdb so we should just go and
run in gdb.

We probably miss a few options to give him more control.