compilers: Fix output ext while compile-only on MSVC
When doing a compile-only self.compile() check, cl.exe expects an output
file of type *.obj, and passing it `output.exe` makes it write the
compiled object to `output.exe.obj`. Detect when we're doing
a compily-only check and change the suffix ourselves.

This is fine most of the time, but fails when you need to inspect the
compiled file manually.

Also store stdout/stderr on the returned object.