has_header: Don't compile, only preprocess
Since we're checking for the existence of a header, just running the
preprocessor is enough. According to my benchmarks, doing this makes the
test roughly 2x faster.

I'm sure this will be useful for other checks too in the future.

This shaves off 7% on the configure time for glib on my machine.

This also fixes the issue that we had earlier where you had to specify
any extra headers needed to resolve symbols in the header being checked
for with `prefix`.