Proponsal for a "query_exists" function for Meson
This function receives a string or an array of strings with paths
(relative or absolute). If at least one of the paths (which can
point to a file or a folder) do exist, it will return a string
with the first one that exists.

If none of those paths do exist, by default it will fail, showing
the list of files not available (but will return an empty string
if the "fail" keyword argument is set to False).

It is possible to exclude files (for checking only folders)
by setting the "check_files" keyword argument to False. Also
it is possible to exclude folders (for checking only files)
by setting the "check_folders" keyword argument to False.

This function can be used to check if libraries that doesn't
have a pkgconfig file are installed, by checking if their
header files exists. Since these can be stored at least at
two possible paths ("/usr/include/..." or "/usr/local/include")
it is mandatory to be able to check every possible place and
fail only if it is not available in any of them.

This function will be very useful for Autovala's Meson backend.